# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: OHB Reinraum Sensor Dashboard — API
  version: 1.0.0
  description: |-
    Schnittstelle des Reinraum-Überwachungssystems.

    **Fehlerformat:** Alle Fehler antworten mit `{ error: { code, message, details? }, requestId }`.
    Die `requestId` steht zusätzlich im Antwort-Header `X-Request-Id` und erlaubt die
    Zuordnung zu den Logzeilen des Servers.

    **Zeitangaben:** durchgängig ISO 8601 mit Zeitzonenangabe (`timestamptz`).

    **Zeitraum-Versionierung:** Zuordnungen und Schwellenwerte werden nie überschrieben.
    Eine Änderung schließt den bisherigen Gültigkeitszeitraum und öffnet einen neuen.
    Auswertungen beziehen sich deshalb immer auf den Stand zum jeweiligen Zeitpunkt.
servers:
  - url: /
    description: Gleiche Herkunft (hinter nginx)
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stabiler Fehlercode für Clients
            message:
              type: string
            details: {}
          required:
            - code
            - message
        requestId:
          type: string
          description: Korrelations-ID, auch im Header X-Request-Id
      required:
        - error
        - requestId
...
    Threshold:
      type: object
      properties:
        id:
          type: integer
        sensor_uuid:
          type: string
        quantity:
          type: string
        min_value:
          type:
            - number
            - "null"
        max_value:
          type:
            - number
            - "null"
        valid_during: {}
      required:
        - id
        - sensor_uuid
        - quantity
        - min_value
        - max_value
    SetThresholdBody:
      type: object
      properties:
        sensor_uuid:
          type: string
          format: uuid
          description: Sensor-Kennung im Format 8-4-4-4-12
        quantity:
          type: string
          minLength: 1
          maxLength: 64
          description: Messgröße
        min_value:
          type:
            - number
            - "null"
        max_value:
          type:
            - number
            - "null"
      required:
        - sensor_uuid
        - quantity
...
    Violation:
      type: object
      properties:
        id:
          type: integer
        sensor_uuid:
          type: string
        sensor_name:
          type:
            - string
            - "null"
        cleanroom_id:
          type:
            - integer
            - "null"
        cleanroom_name:
          type:
            - string
            - "null"
        quantity:
          type: string
        violation_type:
          type: string
          enum:
            - below_min
            - above_max
        threshold_min:
          type:
            - number
            - "null"
        threshold_max:
          type:
            - number
            - "null"
        started_at:
          anyOf:
            - type: string
              format: date-time
            - type: string
        ended_at:
          anyOf:
            - type: string
              format: date-time
            - type: string
            - type: "null"
          description: null = läuft noch
        duration_sec:
          type:
            - number
            - "null"
        first_value:
          type: number
        last_value:
          type:
            - number
            - "null"
        peak_value:
          type: number
          description: Extremster Wert des Ereignisses
        data_points:
          type: integer
        acknowledged:
          type: boolean
      required:
        - id
        - sensor_uuid
        - sensor_name
        - cleanroom_id
        - cleanroom_name
        - quantity
        - violation_type
        - threshold_min
        - threshold_max
        - started_at
        - ended_at
        - duration_sec
        - first_value
        - last_value
        - peak_value
        - data_points
        - acknowledged
    ViolationSummaryEntry:
      type: object
      properties:
        cleanroom_id:
          type:
            - integer
            - "null"
        cleanroom_name:
          type:
            - string
            - "null"
        quantity:
          type: string
        violation_type:
          type: string
          enum:
            - below_min
            - above_max
        event_count:
          type: integer
        total_data_points:
          type: integer
        total_duration_sec:
          type:
            - number
            - "null"
        first_event:
          anyOf:
            - type: string
              format: date-time
            - type: string
        last_event:
          anyOf:
            - type: string
              format: date-time
            - type: string
      required:
        - cleanroom_id
        - cleanroom_name
        - quantity
        - violation_type
        - event_count
        - total_data_points
        - total_duration_sec
        - first_event
        - last_event
...
paths:
...
  /api/thresholds:
    get:
      tags:
        - Schwellenwerte
      summary: Aktuell gültige Schwellenwerte
      parameters:
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: false
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: query
      responses:
        "200":
          description: Schwellenwerte
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Threshold"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Schwellenwerte
      summary: Schwellenwert setzen
      description: Der bisherige Wert wird nicht überschrieben, sondern sein Gültigkeitszeitraum
        abgeschlossen. Auswertungen vergangener Zeiträume bleiben dadurch korrekt.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetThresholdBody"
      responses:
        "201":
          description: Neuer Schwellenwert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threshold"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Konflikt mit dem aktuellen Zustand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/thresholds/{id}:
    delete:
      tags:
        - Schwellenwerte
      summary: Schwellenwert außer Kraft setzen
      description: "Kein echtes Löschen: Der Eintrag bleibt als historischer Beleg erhalten und wird
        lediglich zeitlich abgeschlossen."
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Abgeschlossener Schwellenwert
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  threshold:
                    $ref: "#/components/schemas/Threshold"
                required:
                  - deleted
                  - threshold
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/violations:
    get:
      tags:
        - Alarme
      summary: Schwellenwert-Verletzungen
      description: "Ein Eintrag ist ein EREIGNIS, kein Einzelmesswert: Eine zehnminütige Überschreitung
        bei 2-Sekunden-Takt erzeugt 300 Messwerte, aber genau einen Eintrag mit Extremwert, Dauer
        und Anzahl betroffener Punkte."
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          required: false
          name: cleanroom_id
          in: query
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: false
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 64
            description: Messgröße
          required: false
          description: Messgröße
          name: quantity
          in: query
        - schema:
            type: string
            format: date-time
            description: Zeitpunkt nach ISO 8601
          required: false
          description: Zeitpunkt nach ISO 8601
          name: from
          in: query
        - schema:
            type: string
            format: date-time
            description: Zeitpunkt nach ISO 8601
          required: false
          description: Zeitpunkt nach ISO 8601
          name: to
          in: query
        - schema:
            type: string
            enum:
              - "true"
              - "false"
            description: true = noch laufend
          required: false
          description: true = noch laufend
          name: active
          in: query
        - schema:
            type: string
            enum:
              - "true"
              - "false"
          required: false
          name: acknowledged
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 200
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: Verletzungen, neueste zuerst
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/violations/summary:
    get:
      tags:
        - Alarme
      summary: Aggregat je Reinraum, Messgröße und Verletzungsart
      parameters:
        - schema:
            type: string
            format: date-time
            description: Zeitpunkt nach ISO 8601
          required: false
          description: Zeitpunkt nach ISO 8601
          name: from
          in: query
        - schema:
            type: string
            format: date-time
            description: Zeitpunkt nach ISO 8601
          required: false
          description: Zeitpunkt nach ISO 8601
          name: to
          in: query
      responses:
        "200":
          description: Aggregat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ViolationSummaryEntry"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/violations/{id}/acknowledge:
    patch:
      tags:
        - Alarme
      summary: Verletzung quittieren
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Quittierte Verletzung
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  sensor_uuid:
                    type: string
                  sensor_name:
                    type:
                      - string
                      - "null"
                  cleanroom_id:
                    type:
                      - integer
                      - "null"
                  cleanroom_name:
                    type:
                      - string
                      - "null"
                  quantity:
                    type: string
                  violation_type:
                    type: string
                    enum:
                      - below_min
                      - above_max
                  threshold_min:
                    type:
                      - number
                      - "null"
                  threshold_max:
                    type:
                      - number
                      - "null"
                  started_at:
                    anyOf:
                      - type: string
                        format: date-time
                      - type: string
                  ended_at:
                    anyOf:
                      - type: string
                        format: date-time
                      - type: string
                      - type: "null"
                    description: null = läuft noch
                  duration_sec:
                    type:
                      - number
                      - "null"
                  first_value:
                    type: number
                  last_value:
                    type:
                      - number
                      - "null"
                  peak_value:
                    type: number
                    description: Extremster Wert des Ereignisses
                  data_points:
                    type: integer
                  acknowledged:
                    type: boolean
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...